+++ src/main/java/de/codescape/jira/plugins/scrumpoker/ao/ScrumPokerSession.java
@@ -42,11 +42,23 @@ public interface ScrumPokerSession extends RawEntity<String> {
     void setVisible(boolean visible);
 
     /**
-     * The confirmed vote for this Scrum Poker session if exists.
+     * @deprecated since 4.11 replaced by {@link #getConfirmedEstimate()}
      */
+    @Deprecated
     Integer getConfirmedVote();
 
+    /**
+     * @deprecated since 4.11 replaced by {@link #setConfirmedEstimate(String)}
+     */
+    @Deprecated
     void setConfirmedVote(Integer confirmedVote);
+
+    /**
+     * The confirmed estimate for this Scrum Poker session if exists.
+     */
+    String getConfirmedEstimate();
+
+    void setConfirmedEstimate(String estimate);
 
     /**
      * The unique key of the user who confirmed the estimation.