+++ src/main/java/de/codescape/jira/plugins/scrumpoker/service/EstimateFieldService.java
@@ -1,7 +1,10 @@
 package de.codescape.jira.plugins.scrumpoker.service;
 
 import com.atlassian.activeobjects.tx.Transactional;
+import com.atlassian.jira.issue.Issue;
 import com.atlassian.jira.issue.fields.CustomField;
+
+import java.util.List;
 
 /**
  * Service to access and write the value of the configured estimate field. This is typically the Story Point field
@@ -17,7 +20,7 @@ public interface EstimateFieldService {
      * @param estimate new estimate
      * @return <code>true</code> if value is persisted, otherwise <code>false</code>
      */
-    boolean save(String issueKey, Integer estimate);
+    boolean save(String issueKey, String estimate);
 
     /**
      * Return the estimate custom field.
@@ -25,5 +28,20 @@ public interface EstimateFieldService {
      * @return the estimate custom field
      */
     CustomField findEstimateField();
+
+    /**
+     * Return whether the given issue has the estimate field configured.
+     *
+     * @param issue issue
+     * @return <code>true</code> if estimate field exists, otherwise <code>false</code>
+     */
+    boolean hasEstimateField(Issue issue);
+
+    /**
+     * Return a list of supported custom fields.
+     *
+     * @return list of supported custom fields
+     */
+    List<CustomField> supportedCustomFields();
 
 }